Move common storybook components and assets to storybook-common
