+++ src/hooks/UseArray/Bookshelf/Bookshelf.tsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import { Button } from '../../../common/components'
+import { Button } from '../../../storybook-common/components'
 import useArray from '../useArray'
 import { BOOKS, UNUSED_BOOKS } from './constants'
