+++ src/hooks/UseArray/Bookshelf/constants.ts
@@ -1,4 +1,4 @@
-import { BOOKS_WITH_COLORS } from '../../../common/assets/books'
+import { BOOKS_WITH_COLORS } from '../../../storybook-common/assets/books'
 
 export const BOOKS = BOOKS_WITH_COLORS.slice(0, 15)
 export const UNUSED_BOOKS = BOOKS_WITH_COLORS.slice(15)
