+++ src/storybook-common/assets/books.ts
